Support certificate chain files, @files in ssl_check_chain
The special port 'FILE' now indicates that a chain file
is to be processed instead of contacting a host.  In
common cases, the 'FILE' is optional

The argument form '@file' reads a list of hosts/files to
check from the specified file.  '@file's can be nested.

Also fixes a bug with TLS v1.3