Add "brief" report format for ssl_status
Make the ssl_status reports a bit easier to read by suppressing
duplicate hostnames.  (These occur for dual-certificate hosts when
the certificates expire on the same day.)

The "brief" format is the default, but it may change based on experience.

The --nobrief option delivers the abbreviated report format, which is
more suitable for scripts.